Create the Smith listing test scenario 2
Test Scenario 2:

Bloom Listing Testing Scenarios
Bloom Listing Testing Scenarios
90%
10

The multi-family property has 15 1bedroom units and 2 2 bedroom units. The property manager is looking to open the waitlist for this property. All the units in the property uses the HERA special AMI Chart. The 3 1bedroom units have a maximum income of 60% AMI and the 3 2bedroom units have a maximum income of 50% AMI. All units are reserved for Seniors 65+
Screen reader support enabled.
 
 
 		

The multi-family property has 15 1bedroom units and 2 2 bedroom units. The property manager is looking to open the waitlist for this property. All the units in the property uses the HERA special AMI Chart. The 3 1bedroom units have a maximum income of 60% AMI and the 3 2bedroom units have a maximum income of 50% AMI. All units are reserved for Seniors 65+