Create theartinst_test3.json
Create test case number 3:
The multi-family property has the following units: 5 studio units, 5 1 bedroom units, and 3 2 bedroom units. The property manager is listing three vacancies, one for each unit type, as well as opening the waitlist for the entire property. The studio and 1 bedroom units use the TCAC AMI chart and the 2 bedroom units use the HERA AMI chart. The 5 studio units have a maximum AMI of 40%. For the 1 bedroom unit type, 3 of the units (including the vacant one) have a maximum AMI of 50% while the other 2 units have a maximum AMI of 60%. For the 2 bedroom unit type, 1 of the units (including the vacant one) have a maximum AMI of 60% while the other 2 units have a maximum AMI of 50%. The 2 2bedroom units with 50% AMI are reserved for development disabilities